requirements.txt
flask
flask-cors
qdrant-client
transformers
torch
pillow
python-dotenv
requests
sentencepiece
